Components back in place (otherwise clangd won't work)

+++ stm32_platform/ADC_Readings/platform/inc/pinin.h
@@ -4,9 +4,9 @@
 #include "application_setup.h"
 #include <stddef.h>
 #include <stdint.h>
-uint16_t pinin_temp_sens(void);
-uint16_t pinin_pv(void);
-uint16_t pinin_pv1(void);
+float pinin_temp_sens(void);
+float pinin_pv(void);
+float pinin_pv1(void);
 void pinin_usart(uint8_t *, size_t len, enum HAL_Function_Mode);
 
 #endif /* PININ_H */
